requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
boto3==1.36.19
botocore==1.36.19
cachetools==5.5.1
dnspython==2.7.0
jmespath==1.0.1
motor==3.7.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
s3transfer==0.11.2
six==1.17.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
